+++ resources/views/templates/page/nav/nav-project.blade.php
@@ -1,5 +1,5 @@
-{{-- <x-chief::nav title="Catalogus" inline tagged="nav-catalog"/>
-<x-chief::nav title="Blog" inline tagged="nav-blog"/>
-<x-chief::nav title="Bedrijf" inline tagged="nav-company"/> --}}
+{{--<x-chief::nav title="Catalogus" inline tagged="nav-catalog"/>--}}
+{{--<x-chief::nav title="Blog" inline tagged="nav-blog"/>--}}
+{{--<x-chief::nav title="Bedrijf" inline tagged="nav-company"/>--}}
 
 <x-chief::nav tagged="nav" inline/>