+++ routes/chief-admin-routes.php
@@ -21,6 +21,25 @@ Route::put('links', [\Thinktomorrow\Chief\Site\Urls\Controllers\LinksController:
 Route::delete('remove-redirect/{id}', [\Thinktomorrow\Chief\Site\Urls\Controllers\RemoveRedirectController::class, 'delete'])->name('chief.back.assistants.url.remove-redirect')->where('id', '[0-9]+');
 
 /**
+ * -----------------------------------------------------------------
+ * TAG MANAGEMENT
+ * -----------------------------------------------------------------
+ */
+Route::get('tags', [\Thinktomorrow\Chief\Admin\Tags\Http\TagsController::class,'index'])->name('chief.tags.index');
+Route::post('tags', [\Thinktomorrow\Chief\Admin\Tags\Http\TagsController::class,'store'])->name('chief.tags.store');
+Route::get('tags/create', [\Thinktomorrow\Chief\Admin\Tags\Http\TagsController::class,'create'])->name('chief.tags.create');
+Route::delete('tags/{tag}', [\Thinktomorrow\Chief\Admin\Tags\Http\TagsController::class,'delete'])->name('chief.tags.delete');
+Route::put('tags/{tag}', [\Thinktomorrow\Chief\Admin\Tags\Http\TagsController::class,'update'])->name('chief.tags.update');
+Route::get('tags/{tag}/edit', [\Thinktomorrow\Chief\Admin\Tags\Http\TagsController::class,'edit'])->name('chief.tags.edit');
+
+Route::post('taggroups', [\Thinktomorrow\Chief\Admin\Tags\Http\TagGroupsController::class,'store'])->name('chief.taggroups.store');
+Route::get('taggroups/create', [\Thinktomorrow\Chief\Admin\Tags\Http\TagGroupsController::class,'create'])->name('chief.taggroups.create');
+Route::delete('taggroups/{taggroup}', [\Thinktomorrow\Chief\Admin\Tags\Http\TagGroupsController::class,'delete'])->name('chief.taggroups.delete');
+Route::put('taggroups/{taggroup}', [\Thinktomorrow\Chief\Admin\Tags\Http\TagGroupsController::class,'update'])->name('chief.taggroups.update');
+Route::get('taggroups/{taggroup}/edit', [\Thinktomorrow\Chief\Admin\Tags\Http\TagGroupsController::class,'edit'])->name('chief.taggroups.edit');
+
+
+/**
 * -----------------------------------------------------------------
 * MENU MANAGEMENT
 * -----------------------------------------------------------------