+++ src/Admin/Tags/TagState.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Thinktomorrow\Chief\Admin\Tags;
-
-enum TagState: string
-{
-    case online = 'online';
-    case offline = 'offline';
-}